Trying to add another syntax
